+++ src/Requests/Zones/CreateZone.php
@@ -39,6 +39,6 @@ class CreateZone extends Request
 
     public function createDtoFromResponse(Response $response): Zone
     {
-        return new Zone($response->json('zone'));
+        return Zone::fromArray($response->json('zone'));
     }
 }