+++ src/Requests/Zones/UpdateZone.php
@@ -42,6 +42,6 @@ class UpdateZone extends Request
 
     public function createDtoFromResponse(Response $response): Zone
     {
-        return new Zone($response->json('zone'));
+        return Zone::fromArray($response->json('zone'));
     }
 }